+++ Src/EonSharp/Keystore/KdfTypes.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EonSharp.Keystore
+{
+	public enum KdfTypes
+	{
+		Argon2,
+		Pbkdf2
+
+	}
+}