+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.13.0] - 2018-02-18
 ### Added
 - Filtering to HttpTransportLogger
+- Wallet class
+- Keystore classes and interfaces related to initial implementation of wallet encryption proposal
+- PublicAccountGenerator and PublicKeyPairGenerator classes
 
 ### Fixed
 - BEncoding in AccountRegistration